requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
databricks-sdk>=0.20.0
pydantic>=2.0.0
